requirements.txt
aioboto3==14.0.0
boto3==1.36.11
boto3==1.36.23
botocore==1.36.11
botocore==1.36.23
fastapi==0.115.8
python-dotenv==1.0.1
